[misc] Update install instructions on windows.
