+++ docs/api/gym_jiminy/common/bases/compositions.rst
@@ -1,0 +1,6 @@
+Compositions
+============
+
+.. automodule:: gym_jiminy.common.bases.compositions
+   :members:
+   :show-inheritance:
