+++ docs/api/gym_jiminy/common/bases/index.rst
@@ -6,4 +6,6 @@ Bases
 
    interfaces
    blocks
+   quantities
+   compositions
    pipeline
