+++ docs/api/gym_jiminy/common/bases/quantities.rst
@@ -1,0 +1,6 @@
+Quantities
+==========
+
+.. automodule:: gym_jiminy.common.bases.quantities
+   :members:
+   :show-inheritance:
