--- docs/api/gym_jiminy/common/bases/quantity.rst
@@ -1,6 +1,0 @@
-Quantity
-========
-
-.. automodule:: gym_jiminy.common.bases.quantity
-   :members:
-   :show-inheritance:
