--- docs/api/gym_jiminy/common/bases/reward.rst
@@ -1,6 +1,0 @@
-Reward
-======
-
-.. automodule:: gym_jiminy.common.bases.reward
-   :members:
-   :show-inheritance:
