+++ docs/api/gym_jiminy/common/compositions/index.rst
@@ -1,5 +1,5 @@
-Envs
-====
+Compositions
+============
 
 .. toctree::
    :maxdepth: 1
