+++ docs/api/gym_jiminy/common/quantities/index.rst
@@ -1,5 +1,5 @@
-Envs
-====
+Quantities
+==========
 
 .. toctree::
    :maxdepth: 1
